requirements.txt
pandas == 1.5.3
numpy>=1.23.2
geopandas == 0.13.2
folium == 0.14.0
branca == 0.6.0
seaborn == 0.12.2
matplotlib == 3.7.0
shapely == 2.0.1
hydra-core == 1.3.2 